WS/requirements.txt
Django==5.1.5
django-cors-headers==4.7.0
djangorestframework==3.15.2
psycopg2-binary==2.9.10
pre_commit==4.1.0
python-dotenv==1.0.1
djangorestframework-simplejwt==5.4.0
certifi==2025.4.26
python-certifi-win32==1.6.1
pytest==8.3.5
pytest-django== 4.11.1
